Testing scene maze level
Things affected outside of the scene: creating a duplicate of the red trigger material for green

adding an ontriggercollision to the player controller for the triggers and then an if statement that sets different objects in the scene active or not.

Created scripts for an obstacle in the scene that continuously creates a prefab falling from the sky in an area.